+++ src/derivatives_trading_coin_futures/rest_api/models/query_index_price_constituents_response_constituents_inner.rs
@@ -21,6 +21,10 @@ pub struct QueryIndexPriceConstituentsResponseConstituentsInner {
     pub exchange: Option<String>,
     #[serde(rename = "symbol", skip_serializing_if = "Option::is_none")]
     pub symbol: Option<String>,
+    #[serde(rename = "price", skip_serializing_if = "Option::is_none")]
+    pub price: Option<String>,
+    #[serde(rename = "weight", skip_serializing_if = "Option::is_none")]
+    pub weight: Option<String>,
 }
 
 impl QueryIndexPriceConstituentsResponseConstituentsInner {
@@ -29,6 +33,8 @@ impl QueryIndexPriceConstituentsResponseConstituentsInner {
         QueryIndexPriceConstituentsResponseConstituentsInner {
             exchange: None,
             symbol: None,
+            price: None,
+            weight: None,
         }
     }
 }